refactor: user 패키지 명을 member로 변경
